anuvaad-etl/anuvaad-notifier/requirements.txt
pip==20.1
kafka-python==2.0.1
flask==1.1.2
flask-cors==3.0.8
flask-restful==0.3.8
anuvaad-auditor==0.1.3
requests==2.24.0
Flask-SocketIO==5.0.1
eventlet==0.30.1
